mcp-scan/requirements.txt
agno
firecrawl
fpdf
google-generativeai
python-dotenv
nest_asyncio
colorama
tqdm
gitingest
openai
duckduckgo-search